+++ pkg/kubelet/kubelet_node_status_test.go
@@ -34,6 +34,7 @@ import (
 	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset/fake"
 	"k8s.io/kubernetes/pkg/client/testing/core"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
+	"k8s.io/kubernetes/pkg/kubelet/util/sliceutils"
 	"k8s.io/kubernetes/pkg/runtime"
 	"k8s.io/kubernetes/pkg/util/diff"
 	"k8s.io/kubernetes/pkg/util/rand"
@@ -57,7 +58,7 @@ func generateTestingImageList(count int) ([]kubecontainer.Image, []api.Container
 
 	// expectedImageList is generated by imageList according to size and maxImagesInNodeStatus
 	// 1. sort the imageList by size
-	sort.Sort(byImageSize(imageList))
+	sort.Sort(sliceutils.ByImageSize(imageList))
 	// 2. convert sorted imageList to api.ContainerImage list
 	var expectedImageList []api.ContainerImage
 	for _, kubeImage := range imageList {
