+++ pkg/kubelet/kubelet_test.go
@@ -794,90 +794,6 @@ func TestVolumeUnmountAndDetachControllerEnabled(t *testing.T) {
 	}
 }
 
-func TestPodVolumesExist(t *testing.T) {
-	testKubelet := newTestKubelet(t, false /* controllerAttachDetachEnabled */)
-	kubelet := testKubelet.kubelet
-
-	pods := []*api.Pod{
-		{
-			ObjectMeta: api.ObjectMeta{
-				Name: "pod1",
-				UID:  "pod1uid",
-			},
-			Spec: api.PodSpec{
-				Volumes: []api.Volume{
-					{
-						Name: "vol1",
-						VolumeSource: api.VolumeSource{
-							GCEPersistentDisk: &api.GCEPersistentDiskVolumeSource{
-								PDName: "fake-device1",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			ObjectMeta: api.ObjectMeta{
-				Name: "pod2",
-				UID:  "pod2uid",
-			},
-			Spec: api.PodSpec{
-				Volumes: []api.Volume{
-					{
-						Name: "vol2",
-						VolumeSource: api.VolumeSource{
-							GCEPersistentDisk: &api.GCEPersistentDiskVolumeSource{
-								PDName: "fake-device2",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			ObjectMeta: api.ObjectMeta{
-				Name: "pod3",
-				UID:  "pod3uid",
-			},
-			Spec: api.PodSpec{
-				Volumes: []api.Volume{
-					{
-						Name: "vol3",
-						VolumeSource: api.VolumeSource{
-							GCEPersistentDisk: &api.GCEPersistentDiskVolumeSource{
-								PDName: "fake-device3",
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	stopCh := runVolumeManager(kubelet)
-	defer func() {
-		close(stopCh)
-	}()
-
-	kubelet.podManager.SetPods(pods)
-	for _, pod := range pods {
-		err := kubelet.volumeManager.WaitForAttachAndMount(pod)
-		if err != nil {
-			t.Errorf("Expected success: %v", err)
-		}
-	}
-
-	for _, pod := range pods {
-		podVolumesExist := kubelet.podVolumesExist(pod.UID)
-		if !podVolumesExist {
-			t.Errorf(
-				"Expected to find volumes for pod %q, but podVolumesExist returned false",
-				pod.UID)
-		}
-	}
-}
-
 type stubVolume struct {
 	path string
 	volume.MetricsNil
