+++ pkg/kubelet/util.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2014 The Kubernetes Authors.
+Copyright 2016 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -128,3 +128,6 @@ func dirExists(path string) bool {
 	}
 	return s.IsDir()
 }
+
+// empty is a placeholder type used to implement a set
+type empty struct{}
